requirements.txt
pulumi>=3
pulumi-gcp
requests
python-dotenv